stream-elt/requirements.txt
confluent-kafka==2.3.0
websocket-client==1.8.0
python-dotenv==1.0.1
requests==2.32.3
fastavro==1.9.5
